+++ src/main/java/io/xlorey/fluxloader/client/core/Core.java
@@ -7,7 +7,6 @@ import io.xlorey.fluxloader.client.handlers.OnPostUIDrawHandler;
 import io.xlorey.fluxloader.enums.EventPriority;
 import io.xlorey.fluxloader.plugin.PluginManager;
 import io.xlorey.fluxloader.shared.EventManager;
-import io.xlorey.fluxloader.utils.Logger;
 
 /**
  * Author: Deknil
@@ -23,14 +22,11 @@ public class Core {
     public static boolean isPluginsExecuted = false;
 
     /**
-     * Initializing the loader
+     * Initializing the loader. Called after loading mods and basic initializations,
+     * namely at the moment after the Translator is initialized
      * @exception Exception in cases of unsuccessful core initialization
      */
     public static void init() throws Exception {
-        Logger.printCredits();
-
-        Logger.print("FluxLoader Core initialization for the client...");
-
         EventManager.subscribe(new OnInitWorldHandler(), EventPriority.HIGHEST);
         EventManager.subscribe(new OnMainMenuEnterHandler(), EventPriority.HIGHEST);
         EventManager.subscribe(new OnPostTickRenderThreadHandler(), EventPriority.HIGHEST);