+++ src/main/java/io/xlorey/fluxloader/patches/PatchGameWindow.java
@@ -4,6 +4,7 @@ import io.xlorey.fluxloader.client.core.Core;
 import io.xlorey.fluxloader.client.core.StateManager;
 import io.xlorey.fluxloader.shared.EventManager;
 import io.xlorey.fluxloader.utils.Constants;
+import io.xlorey.fluxloader.utils.Logger;
 import io.xlorey.fluxloader.utils.PatchTools;
 import javassist.CannotCompileException;
 import javassist.expr.ExprEditor;
@@ -33,7 +34,21 @@ public class PatchGameWindow extends PatchFile{
         // Basic initialization of FluxLoader
         PatchTools.patchMethod(getClassName(), "init", (ctClass, ctMethod) -> {
             try {
-                ctMethod.insertBefore("{ DoLoadingText(\"Loading Flux Loader\"); "+Core.class.getName() + ".init(); }");
+                ctMethod.insertBefore("{" +
+                        Logger.class.getName() + ".printCredits();" +
+                        Logger.class.getName() + ".print(\"FluxLoader Core initialization for the client...\");" +
+                        "}");
+                ctMethod.instrument(new ExprEditor() {
+                    public void edit(MethodCall m) throws CannotCompileException {
+                        if (m.getClassName().contains("Translator") && m.getMethodName().equals("loadFiles")) {
+                            String code =  "DoLoadingText(\"Loading Flux Loader\"); " + Core.class.getName() + ".init();";
+                            m.replace("{" +
+                                    "$proceed($$);" +
+                                    code +
+                                    "}");
+                        }
+                    }
+                });
                 ctMethod.insertAfter(StateManager.class.getName() + ".init();");
                 ctMethod.insertAfter(EventManager.class.getName() + ".invokeEvent(\"onGameWindowInitialized\", new Object[0]);");
             } catch (CannotCompileException e) {