+++ changelog.md
@@ -2,8 +2,12 @@
 - Added `onLuaFilesLoaded` and `onLuaScriptExecute` events
 - Added `onChangeLanguage` events
 - Added translation system to `TranslationManager`
+- Added logic for extracting lua files to the plugin configuration folder
+- Added methods for working with lua
+- The initialization location of the client and server core has been changed
 - Fixed names of methods for exposing classes and objects
 - Fixed log output when creating backups and restoring
+- Minor fixes in config logic
 
 # FluxLoader v0.9.0
 - Added methods for getting and setting lists of specific types for plugin configs